backend/requirements.txt
# Backend requirements
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
pymongo==4.6.0
motor==3.3.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
python-dotenv==1.0.0
redis==5.0.1
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
matplotlib==3.7.1
seaborn==0.12.2
joblib==1.3.2
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
slowapi==0.1.9
python-jose==3.3.0
bcrypt==4.1.2
